// lib/interface.ts
export interface simpleBlogCard {
  title: string;
  slugCurrent: string;
  categoryTitle: any;
  imagePost: any;
  author: any;
  description: string;
  updatedAt: string;
}
export interface fullBlog {
  slugCurrent: string;
  title: string;
  categoryTitle: any;
  imagePost: any;
  author: any;
  publishedAt: string;
  contentPost: any;
}